+++ MawCluster/Probability.pyx
@@ -38,8 +38,8 @@ PROTOCOLS = ["ffpe", "amplicon", "cf", "other"]
 
 
 @memoize
-@cython.locals(DOC=cython.long, pVal=np.longdouble_t,
-               AC=cython.long)
+@cython.locals(DOC=cython.int, pVal=np.longdouble_t,
+               AC=cython.int)
 def ConfidenceIntervalAAF(AC, DOC, pVal=defaultPValue,
                           method="agresti_coull"):
     """
@@ -58,8 +58,8 @@ def ConfidenceIntervalAAF(AC, DOC, pVal=defaultPValue,
 
 
 @cython.returns(np.ndarray)
-def ConfidenceIntervalAI(cython.long Allele1,
-                         cython.long Allele2, np.longdouble_t pVal=defaultPValue,
+def ConfidenceIntervalAI(cython.int Allele1,
+                         cython.int Allele2, np.longdouble_t pVal=defaultPValue,
                          cython.str method="agresti_coull"):
     """
     Returns the confidence interval for an allelic imbalance
@@ -86,8 +86,8 @@ def ConfidenceIntervalAI(cython.long Allele1,
 
 
 @cython.returns(tuple)
-def MakeAICall(cython.long Allele1,
-               cython.long Allele2,
+def MakeAICall(cython.int Allele1,
+               cython.int Allele2,
                np.longdouble_t pVal=defaultPValue,
                cython.str method="agresti_coull"):
     """
@@ -107,9 +107,9 @@ def MakeAICall(cython.long Allele1,
 
 
 @memoize
-@cython.locals(n=cython.long, p=np.longdouble_t,
+@cython.locals(n=cython.int, p=np.longdouble_t,
                pVal=np.longdouble_t)
-@cython.returns(cython.long)
+@cython.returns(cython.int)
 def GetCeiling(n, p=0.0, pVal=defaultPValue):
     """
     Returns the maximum fraction of events per sample with a p value of pVal,
@@ -144,7 +144,7 @@ def StirlingsFact(n, k):
 
 
 @memoize
-@cython.locals(p=np.longdouble_t, k=cython.long, n=cython.long)
+@cython.locals(p=np.longdouble_t, k=cython.int, n=cython.int)
 @cython.returns(np.longdouble_t)
 def SamplingFrac(n, p=0., k=1):
     """
@@ -159,7 +159,7 @@ def SamplingFrac(n, p=0., k=1):
 
 
 @memoize
-@cython.locals(p=np.longdouble_t, k=cython.long, n=cython.long)
+@cython.locals(p=np.longdouble_t, k=cython.int, n=cython.int)
 @cython.returns(np.longdouble_t)
 def SamplingFrac_(n, p=0., k=1):
     """
@@ -173,7 +173,7 @@ def SamplingFrac_(n, p=0., k=1):
 
 
 @memoize
-@cython.locals(p=np.longdouble_t, k=cython.long, n=cython.long)
+@cython.locals(p=np.longdouble_t, k=cython.int, n=cython.int)
 @cython.returns(np.ndarray)
 def GetUnscaledProbs(n, p=0.):
     """
@@ -186,7 +186,7 @@ def GetUnscaledProbs(n, p=0.):
 
 
 @memoize
-@cython.locals(p=np.longdouble_t, k=cython.long, n=cython.long)
+@cython.locals(p=np.longdouble_t, k=cython.int, n=cython.int)
 @cython.returns(np.ndarray)
 def GetUnscaledProbs_(n, p=0.):
     """
@@ -199,21 +199,21 @@ def GetUnscaledProbs_(n, p=0.):
 
 
 @memoize
-@cython.locals(p=np.longdouble_t, k=cython.long, n=cython.long)
+@cython.locals(p=np.longdouble_t, k=cython.int, n=cython.int)
 @cython.returns(np.longdouble_t)
 def PartitionFunction(n, p=0.1):
     return nsum(GetUnscaledProbs(n, p=p), 0)
 
 
 @memoize
-@cython.locals(p=np.longdouble_t, k=cython.long, n=cython.long)
+@cython.locals(p=np.longdouble_t, k=cython.int, n=cython.int)
 @cython.returns(np.longdouble_t)
 def PartitionFunction_(n, p=0.1):
     return nsum(GetUnscaledProbs_(n, p=p), 0)
 
 
 @memoize
-@cython.locals(p=np.longdouble_t, n=cython.long, k=cython.long)
+@cython.locals(p=np.longdouble_t, n=cython.int, k=cython.int)
 @cython.returns(np.ndarray)
 def SamplingProbDist(n, p=0.):
     """
@@ -230,7 +230,7 @@ def SamplingProbDist(n, p=0.):
 
 
 @memoize
-@cython.locals(p=np.longdouble_t, n=cython.long, k=cython.long)
+@cython.locals(p=np.longdouble_t, n=cython.int, k=cython.int)
 @cython.returns(np.ndarray)
 def SamplingProbDist_(n, p=0.):
     """
@@ -247,7 +247,7 @@ def SamplingProbDist_(n, p=0.):
 
 
 @memoize
-@cython.locals(p=np.longdouble_t, n=cython.long, k=cython.long,
+@cython.locals(p=np.longdouble_t, n=cython.int, k=cython.int,
                PartitionFn=np.longdouble_t)
 def SamplingProbMoments(n, p=0.):
     """
@@ -262,7 +262,7 @@ def SamplingProbMoments(n, p=0.):
 
 
 @memoize
-@cython.locals(p=np.longdouble_t, n=cython.long, k=cython.long,
+@cython.locals(p=np.longdouble_t, n=cython.int, k=cython.int,
                PartitionFn=np.longdouble_t)
 def SamplingProbMoments_(n, p=0.):
     """
